refactor: move osc user bootstrap helper
